refactor:  rename file to match multivariate
