+++ copulax/_src/univariate/_normal_mixture.py
